move ResourceForUtilsCache to internal package
